[BACKLOG-35933] - Move Infobright from core to plugin changes
