+++ assemblies/plugins/pom.xml
@@ -30,6 +30,7 @@
     <ms-access-plugin.version>${project.version}</ms-access-plugin.version>
     <oracle-bulk-loader-plugin.version>${project.version}</oracle-bulk-loader-plugin.version>
     <postgresql-db-bulk-loader-plugin.version>${project.version}</postgresql-db-bulk-loader-plugin.version>
+    <infobright-bulk-loader-plugin.version>${project.version}</infobright-bulk-loader-plugin.version>
     <lucid-db-streaming-loader-plugin.version>${project.version}</lucid-db-streaming-loader-plugin.version>
     <gp-bulk-loader-plugin.version>${project.version}</gp-bulk-loader-plugin.version>
     <pentaho-googledrive-vfs-plugin.version>${project.version}</pentaho-googledrive-vfs-plugin.version>
@@ -208,6 +209,18 @@
       <groupId>org.pentaho.di.plugins</groupId>
       <artifactId>postgresql-db-bulk-loader-plugins</artifactId>
       <version>${postgresql-db-bulk-loader-plugin.version}</version>
+      <type>zip</type>
+      <exclusions>
+        <exclusion>
+          <groupId>*</groupId>
+          <artifactId>*</artifactId>
+        </exclusion>
+      </exclusions>
+    </dependency>
+    <dependency>
+      <groupId>org.pentaho.di.plugins</groupId>
+      <artifactId>infobright-bulk-loader-plugins</artifactId>
+      <version>${infobright-bulk-loader-plugin.version}</version>
       <type>zip</type>
       <exclusions>
         <exclusion>
