+++ engine/src/main/resources/kettle-steps.xml
@@ -12,7 +12,6 @@
  <step id="GetSlaveSequence"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.GetSlaveSequence</description> <classname>org.pentaho.di.trans.steps.getslavesequence.GetSlaveSequenceMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Transform</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.GetSlavesequence</tooltip> <iconfile>ui/images/SEQ.svg</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Get+ID+from+Slave+Server</documentation_url> <cases_url/> <forum_url/> </step>
  <step id="TableOutput"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.Output</description> <classname>org.pentaho.di.trans.steps.tableoutput.TableOutputMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Output</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.TableOutput</tooltip> <iconfile>ui/images/TOP.svg</iconfile> <documentation_url>Products/Table_Output</documentation_url> <cases_url/> <forum_url/> </step>
  <step id="TableAgileMart"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.TableAgileMart</description> <classname>org.pentaho.di.trans.steps.tableagilemart.TableAgileMartMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Agile</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.TableAgileMart</tooltip> <iconfile>ui/images/TAM.svg</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Table+Agile+Mart</documentation_url> <cases_url/> <forum_url/> </step>
- <step id="InfobrightOutput"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.InfobrightOutput</description> <classname>org.pentaho.di.trans.steps.infobrightoutput.InfobrightLoaderMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Bulk</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.InfobrightOutput</tooltip> <iconfile>ui/images/BLKIB.svg</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Infobright+Loader</documentation_url> <cases_url/> <forum_url/> </step>
  <step id="SortRows"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.SortRows</description> <classname>org.pentaho.di.trans.steps.sort.SortRowsMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Transform</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.Sortrows</tooltip> <iconfile>ui/images/SRT.svg</iconfile> <documentation_url>Products/Sort_rows</documentation_url> <cases_url/> <forum_url/> </step>
  <step id="Injector"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.Injector</description> <classname>org.pentaho.di.trans.steps.injector.InjectorMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Inline</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.Injector</tooltip> <iconfile>ui/images/INJ.svg</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Injector</documentation_url> <cases_url/> <forum_url/> </step>
  <step id="FilterRows"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.FilterRows</description> <classname>org.pentaho.di.trans.steps.filterrows.FilterRowsMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Flow</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.FilterRows</tooltip> <iconfile>ui/images/FLT.svg</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Filter+rows</documentation_url> <cases_url/> <forum_url/> </step>
