+++ engine/src/main/resources/org/pentaho/di/trans/step/messages/messages_en_US.properties
@@ -313,7 +313,6 @@ BaseStep.TypeLongDesc.TextFileOutputLegacy=Text file output
 BaseStep.TypeTooltipDesc.Rowflattener=Flattens consecutive rows based on the order in which they appear in the input stream
 BaseStep.TypeLongDesc.Output=Table output
 BaseStep.TypeLongDesc.TableAgileMart=Table Agile Mart
-BaseStep.TypeLongDesc.InfobrightOutput=Infobright loader
 BaseStep.TypeLongDesc.XSDValidator=XSD validator
 BaseStep.Category.Job=Job
 BaseStep.TypeLongDesc.UniqueRows=Unique rows
