--- engine/src/main/resources/org/pentaho/di/trans/steps/infobrightoutput/messages/messages_en_US.properties
@@ -1,53 +1,0 @@
-#####################################################################
-##
-##  InfobrightLoaderDialog
-##
-#####################################################################
-InfobrightLoaderDialog.Stepname.Label=Step name 
-InfobrightLoaderDialog.Shell.Title=Infobright loader
-InfobrightLoaderDialog.Connection.Label=Database connection
-
-InfobrightLoaderDialog.TargetSchema.Label=Target schema
-InfobrightLoaderDialog.TargetTable.Label=Target table
-InfobrightLoaderDialog.RejectErrors.Label=Reject errors
-InfobrightLoaderDialog.Dataformat.Label=Infobright product
-InfobrightLoaderDialog.Charset.Label=Character set
-InfobrightLoaderDialog.AgentPort.Label=Agent port
-InfobrightLoaderDialog.DebugFile.Label=Debug file
-
-InfobrightLoaderDialog.Illegal.Dialog.Settings.Title=Illegal value
-InfobrightLoaderDialog.Illegal.Dialog.Settings.Message=Illegal value
-
-#####################################################################
-##
-##  InfobrightLoaderStep
-##
-#####################################################################
-InfobrightLoader.Log.StartingToRun=Starting to run...
-InfobrightLoader.Log.LineNumber=Line number
-InfobrightLoader.Log.UnexpectedError=Unexpected error in ''
-InfobrightLoader.Log.FailedToKillQuery=Failed to kill loader query
-
-#####################################################################
-##
-##  InfobrightLoaderMetadata
-##
-#####################################################################
-InfobrightLoaderMeta.CheckResult.ConnectionExists = Connection exists
-InfobrightLoaderMeta.CheckResult.ConnectionOk = Connection to database OK
-InfobrightLoaderMeta.CheckResult.TableAccessible = Table [{0}] exists and is accessible
-InfobrightLoaderMeta.CheckResult.TableOk = Table [{0}] is readable and we received the fields from it.
-InfobrightLoaderMeta.CheckResult.FieldsReceived = Step is connected to previous one, receiving {0} fields
-InfobrightLoaderMeta.CheckResult.FieldsNotFoundInOutput = Fields in input stream, not found in output table:\n\n{0}
-InfobrightLoaderMeta.CheckResult.AllFieldsFoundInOutput = All fields, coming from previous steps, are found in the output table
-InfobrightLoaderMeta.CheckResult.FieldsNotFound = Fields in table, not found in input stream:\n\n{0}
-InfobrightLoaderMeta.CheckResult.AllFieldsFound = All fields in the table are found in the input stream, coming from previous steps
-InfobrightLoaderMeta.CheckResult.NoFields = Couldn''t find fields from previous steps, check the hops...\!
-InfobrightLoaderMeta.CheckResult.TableNotAccessible = Couldn''t read the table info, please check the table-name & permissions.
-InfobrightLoaderMeta.CheckResult.TableError = Table [{0}] doesn''t exist or can''t be read on this database connection.
-InfobrightLoaderMeta.CheckResult.NoTableName = No table name was entered in this step.
-InfobrightLoaderMeta.CheckResult.UndefinedError = An error occurred: {0}
-InfobrightLoaderMeta.CheckResult.NoConnection = Please select or create a connection to use
-InfobrightLoaderMeta.CheckResult.ExcpectedInputOk=Step is receiving info from other steps.
-InfobrightLoaderMeta.CheckResult.MissingExpectedInput=No input received from other steps\!
-InfobrightLoaderMeta.GetSQL.NoConnectionDefined=There is no connection defined in this step.
