+++ plugins/infobright-bulk-loader/assemblies/pom.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
+         xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <groupId>org.pentaho.di.plugins</groupId>
+    <artifactId>infobright-bulk-loader</artifactId>
+    <version>9.4.0.0-SNAPSHOT</version>
+  </parent>
+
+  <artifactId>infobright-bulk-loader-assemblies</artifactId>
+  <version>9.4.0.0-SNAPSHOT</version>
+  <packaging>pom</packaging>
+
+  <name>PDI Infobright Bulk Loader Plugins Assemblies</name>
+
+  <modules>
+    <module>plugin</module>
+  </modules>
+</project>
