+++ plugins/infobright-bulk-loader/impl/src/main/resources/org/pentaho/di/trans/steps/infobrightoutput/messages/messages_fr_FR.properties
