+++ plugins/infobright-bulk-loader/impl/src/main/resources/org/pentaho/di/trans/steps/infobrightoutput/messages/messages_ja_JP.properties
