+++ plugins/infobright-bulk-loader/ui/src/main/resources/org/pentaho/ui/database/infobright_jndi.xul
@@ -1,0 +1,8 @@
+				<groupbox id="database-options-box" orient="vertical" flex="1"
+					xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
+					xmlns:pen="http://www.pentaho.org/2008/xul" >
+					<caption label="${DatabaseDialog.USER_SETTINGS}" />
+					<label id="database-name-label" value="${DatabaseDialog.USER_JNDI_NAME}" />
+					<textbox pen:customclass="variabletextbox" id="database-name-text" />
+					<checkbox id="result-streaming-check" label="${DatabaseDialog.label.MySQLStreamResults}" checked="false"/>					
+				</groupbox>
