+++ plugins/pom.xml
@@ -76,6 +76,7 @@
         <module>ms-access</module>
         <module>postgresql-db-bulk-loader</module>
         <module>oracle-bulk-loader</module>
+        <module>infobright-bulk-loader</module>
         <module>xml</module>
         <module>streaming</module>
         <module>metastore-locator</module>
